feature: add 23 entities

+++ database/ddl/002_rol.sql
@@ -1,5 +1,6 @@
+-- 1. Catálogos básicos, rol, marca, banco, etc. 
 CREATE TABLE rol (
-    id_rol SMALLINT PRIMARY KEY,
+    id_rol SERIAL PRIMARY KEY,
     nombre VARCHAR(30) NOT NULL UNIQUE,
     descripcion VARCHAR(30)
 );
